requirements.txt
langchain==0.0.266
streamlit
openai
unstructured==0.10.2
python-dotenv